Return the global token for the read path if they're on dnixd
